+++ crates/cubecl-runtime/src/memory_management/base.rs
@@ -3,21 +3,21 @@
 /// wasted in total.
 pub struct MemoryUsage {
     /// The number of allocations currently active.
-    pub number_allocs: usize,
+    pub number_allocs: u64,
     /// The number of bytes that are currently actually in use.
     ///
     /// This doesn't include any padding or other memory that needs to be
     /// reserved, and is the minimum amount of memory that could possible
     /// be allocated.
-    pub bytes_in_use: usize,
+    pub bytes_in_use: u64,
     /// The amount of bytes used for padding memory in currently active allocations.
-    pub bytes_padding: usize,
+    pub bytes_padding: u64,
     /// The total amount of memory reserved on the device.
     ///
     /// This will be at least as much as bytes_in_use but in practice will
     /// be higher, as allocations reserve memory for future allocations
     /// and for padding.
-    pub bytes_reserved: usize,
+    pub bytes_reserved: u64,
 }
 
 impl MemoryUsage {
@@ -31,7 +31,7 @@ impl MemoryUsage {
     }
 }
 
-fn bytes_format(bytes: usize) -> String {
+fn bytes_format(bytes: u64) -> String {
     let unit = 1000;
 
     if bytes < unit {
